Compute averages per hour array and them as a response.
This commit add several methods to actually request InfluxDB in order to
retrieve averages per hour as an array. Now when the webserver receive a GET
request on its /api/1/pings/:origin/hours endpoint, it sends as a response the
JSON array of the average `tranfer_time_ms` per hour since the oldest timestamp
until now.